Rewrite all tests to use jest
- Remove unused packages
- Remove AVA configs from all packages
- Add information to run tests in README.md
