+++ .prettierignore
@@ -1,0 +1,1 @@
+package.json
