+++ README.md
@@ -12,8 +12,16 @@ Mono-repo for npm packages based on Lerna.
 - [composite-logger](packages/composite-logger)
 - [composite-observer](packages/composite-observer)
 
-
 TODO:
 * Add BreakpointManager (Singleton)
 * Add PubSub (Singleton)
 * Add Scrollhandler (Singleton)
+
+Running a test for a specifig package only: 
+```bash
+# Run the tests once
+$ npm run test --testMatch ./packages/<packagename>/**/*.test.js
+
+# Run the tests in watch mode
+$ npm run test-watch --testMatch ./packages/<packagename>/**/*.test.js
+```
