+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "javascript-packages",
   "devDependencies": {
-    "ava": "1.0.0-beta.3",
     "babel-cli": "6.26.0",
     "babel-loader": "7.1.3",
     "babel-plugin-external-helpers": "6.22.0",
@@ -16,16 +15,16 @@
     "jest": "22.4.2",
     "lerna": "2.9.0",
     "lint-staged": "7.0.0",
-    "mock-local-storage": "1.0.5",
     "prettier": "1.10.2",
-    "sinon": "4.4.2",
     "webpack": "4.0.0",
     "webpack-cli": "2.0.9"
   },
   "scripts": {
-    "lint": "eslint --fix \"{,**,**/**}*.js\"",
+    "lint": "lint-staged",
+    "lint-all": "eslint --fix \"{,**,**/**}*.js\"",
     "build": "lerna run build",
-    "test": "ava && jest --silent",
+    "test": "jest --silent",
+    "test-watch": "jest --watchAll",
     "update-dependencies": "npx npm-check-updates -u -a",
     "precommit": "npm run lint && npm test && npm run build",
     "bootstrap": "lerna bootstrap"
@@ -42,23 +41,8 @@
   },
   "jest": {
     "testMatch": [
-      "**/?(*.)(test).js?(x)"
-    ]
-  },
-  "ava": {
-    "files": [
-      "packages/**/*.spec.js",
-      "!**/node_modules/"
-    ],
-    "sources": [
-      "packages/**/*.js",
-      "!**/node_modules/"
-    ],
-    "concurrency": 5,
-    "failFast": true,
-    "powerAssert": true,
-    "require": [
-      "babel-register"
+      "**/?(*.)(test).js?(x)",
+      "**/?(*.)(spec).js?(x)"
     ]
   }
 }
