+++ packages/composite-observer/package.json
@@ -3,8 +3,6 @@
   "version": "0.0.10",
   "description": "Simple observer composite to use for your factories",
   "scripts": {
-    "start": "npm run dev",
-    "dev": "lerna exec --scope=@unic/composite-observer -- ava --watch",
     "build": "lerna exec --scope=@unic/composite-observer -- webpack --mode production --module-bind js=babel-loader"
   },
   "main": "dist/main.js",
@@ -18,19 +16,5 @@
   "license": "Apache-2.0",
   "publishConfig": {
     "access": "public"
-  },
-  "ava": {
-    "files": [
-      "src/**/*.spec.js"
-    ],
-    "sources": [
-      "src/**/*.js"
-    ],
-    "concurrency": 5,
-    "failFast": true,
-    "powerAssert": true,
-    "require": [
-      "babel-register"
-    ]
   }
 }
