+++ packages/composite-observer/src/index.test.js
@@ -1,0 +1,7 @@
+import observer from './index';
+
+test('The right methods are exposed', () => {
+  expect(['on', 'off', 'trigger']).toEqual(
+    expect.arrayContaining(Object.keys(observer())),
+  );
+});
